backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
PyMySQL
python-dotenv
python-decouple
channels
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl
celery
django-celery-beat
django-celery-results
redis
